requirements.txt
pyrogram
python-dotenv
Flask
gunicorn
aiohttp